Remove MZGLKitViewController.h
